【Hackathon 8th No.10】DrivAerNet 论文复现 (#1068)
* Create ReadME.md

* Add files via upload

* Delete examples/fsi/DrivAerNet directory

* Create ReadME.md

* Add files via upload

* drivaernet

* Delete examples/fsi/DrivAerNet-paddle-convert-main directory

* modify some style

* modify yaml

* update example drivaernet code

* update requirments and delete fsi/DrivAerNet.py etc.

* Update drivaernet_dataset.py

* update drivaernet_dataset.py

* update drivaernet_dataset.py

* Update __init__.py

* Update __init__.py

* Update DrivAerNet.py

* Update DrivAerNet.py

* Update __init__.py

* update

* support fraction of the training data

* modify some error about version

* Update regdgcnn.py

* Delete examples/DrivAerNet/requirments.txt

* Delete docs/zh/examples/drivaernet directory

* Update DriveAerNet.yaml

* Rename DrivAerNet.md to drivaernet.md

* Update mkdocs.yml

* Update drivaernet.md

* Update DriveAerNet.yaml

* Update DrivAerNet.py

* Update DriveAerNet.yaml

* Update DrivAerNet.py

* Update DrivAerNet.py

* Update drivaernet.md

* Update DrivAerNet.py

* Update DriveAerNet.yaml

* Update drivaernet_dataset.py

* Update solver.py

* update metric.md

* Update metric.md

* Update optimizer.md

* Update arch.md

* Update DrivAerNetDataset dataset.md

* Update drivaernet_dataset.py

* Update drivaernet.md

* Update dataset.md

* Update arch.md

* Update lr_scheduler.py

* Delete docs/zh/api/arch.md

* Delete docs/zh/api/data/dataset.md

* Delete ppsci/optimizer/lr_scheduler.py

* Update optimizer.md

* Delete docs/zh/api/optimizer.md

* Create arch.md

* Create dataset.md

* Create optimizer.md

* Update lr_scheduler.md

* Create lr_scheduler.py

* Update solver.py

* Update drivaernet.md

* Update arch.md

* Update dataset.md

* Update lr_scheduler.py

* Update lr_scheduler.md

* Rename DriveAerNet.yaml to driveaernet.yaml

* Update and rename DrivAerNet.py to drivaernet.py

* Rename driveaernet.yaml to drivaernet.yaml

* Update drivaernet.md

* Update r2_score.py

* Update max_ae.py

* Update drivaernet_dataset.py

* Update drivaernet.md

* Update solver.py

* Update solver.py

* Update drivaernet.py

* Create drivaernet

* Delete examples/drivaernet

* Create drivaernet

* Delete examples/drivaernet

* Delete examples/DrivAerNet directory

* Create drivaernet.py

* Create drivaernet.yaml

* Update drivaernet.yaml

* Update drivaernet.md

* Update drivaernet.md

* Update regdgcnn.py

* Update lr_scheduler.py

* Update lr_scheduler.py

* Update solver.py

* Update lr_scheduler.py

* Update examples/drivaernet/drivaernet.py


* Update examples/drivaernet/conf/drivaernet.yaml


* Update examples/drivaernet/conf/drivaernet.yaml


* Update examples/drivaernet/conf/drivaernet.yaml


* Update examples/drivaernet/drivaernet.py


* Update ppsci/arch/regdgcnn.py


* Update ppsci/arch/regdgcnn.py


* Update regdgcnn.py

* Update lr_scheduler.py

* Update drivaernet_dataset.py

* Update drivaernet_dataset.py

* Update solver.py

* Update ppsci/solver/solver.py

* modify markdown and dataset_load class file

* move the class of apply_augmention into drivaernet_dataset.py and modify the markdown file about apply_augmention

* remove left bracket and right bracket in data_cfg

---------
