+++ examples/drivaernet/drivaernet.py
@@ -35,9 +35,9 @@ def train(cfg: DictConfig):
         "dataset": {
             "name": "DrivAerNetDataset",
             "root_dir": cfg.ARGS.dataset_path,
-            "input_keys": ("vertices",),
-            "label_keys": ("cd_value",),
-            "weight_keys": ("weight_keys",),
+            "input_keys": cfg.MODEL.input_keys,
+            "label_keys": cfg.MODEL.output_keys,
+            "weight_keys": cfg.MODEL.weight_keys,
             "subset_dir": cfg.ARGS.subset_dir,
             "ids_file": cfg.TRAIN.train_ids_file,
             "csv_file": cfg.ARGS.aero_coeff,
@@ -61,9 +61,9 @@ def train(cfg: DictConfig):
         "dataset": {
             "name": "DrivAerNetDataset",
             "root_dir": cfg.ARGS.dataset_path,
-            "input_keys": ("vertices",),
-            "label_keys": ("cd_value",),
-            "weight_keys": ("weight_keys",),
+            "input_keys": cfg.MODEL.input_keys,
+            "label_keys": cfg.MODEL.output_keys,
+            "weight_keys": cfg.MODEL.weight_keys,
             "subset_dir": cfg.ARGS.subset_dir,
             "ids_file": cfg.TRAIN.eval_ids_file,
             "csv_file": cfg.ARGS.aero_coeff,
@@ -149,9 +149,9 @@ def evaluate(cfg: DictConfig):
         "dataset": {
             "name": "DrivAerNetDataset",
             "root_dir": cfg.ARGS.dataset_path,
-            "input_keys": ("vertices",),
-            "label_keys": ("cd_value",),
-            "weight_keys": ("weight_keys",),
+            "input_keys": cfg.MODEL.input_keys,
+            "label_keys": cfg.MODEL.output_keys,
+            "weight_keys": cfg.MODEL.weight_keys,
             "subset_dir": cfg.ARGS.subset_dir,
             "ids_file": cfg.EVAL.ids_file,
             "csv_file": cfg.ARGS.aero_coeff,
